+++ src/components/common/Card/Card.module.scss
