+++ src/components/common/KebabButton/KebabButton.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import styles from "./KebabButton.module.scss";
 import Image from "next/image";
 import { CardProps, FolderInterface, ModalInterface } from "@/types";
-import { DeleteCardModal, AddCardModal } from "../modals";
+import { DeleteCardModal, AddCardModal } from "@/components/modals";
 interface Props extends CardProps {
   onClick: (m: ModalInterface) => void;
   folderList?: FolderInterface[];
