+++ src/components/common/Spinner/Spinner.module.scss
