+++ src/components/common/index.tsx
@@ -1,0 +1,9 @@
+import Card from "./Card/Card";
+import CardList from "./CardList/CardList";
+import Footer from "./Footer/Footer";
+import ResponsiveImg from "./ResponsiveImg/ResponsiveImg";
+import SearchBar from "./SearchBar/SearchBar";
+import Spinner from "./Spinner/Spinner";
+import Nav from "./Nav/Nav";
+
+export { Card, CardList, Footer, ResponsiveImg, SearchBar, Spinner, Nav };
