--- src/components/index.ts
@@ -1,8 +1,0 @@
-import Nav from "./Nav/Nav";
-import Footer from "./Footer/Footer";
-import Card from "./Card/Card";
-import CardList from "./CardList/CardList";
-import SearchBar from "./SearchBar/SearchBar";
-import SnsSign from "./SnsSign/SnsSign";
-
-export { Nav, Footer, Card, CardList, SearchBar, SnsSign };
