+++ src/components/shared/ShareHeader/ShareHeader.module.scss
