+++ src/components/shared/index.tsx
@@ -1,0 +1,3 @@
+import ShareHeader from "./ShareHeader/ShareHeader";
+
+export { ShareHeader };
