+++ src/components/sign/index.tsx
@@ -1,0 +1,4 @@
+import Input from "./Input/Input";
+import SnsSign from "./SnsSign/SnsSign";
+
+export { Input, SnsSign };
