+++ src/hooks/index.tsx
@@ -1,4 +1,3 @@
-import useKeywordInput from "./useKeywordInput";
 import useOpenModal from "./useOpenModal";
 
-export { useKeywordInput, useOpenModal };
+export { useOpenModal };
