+++ src/pages/signin/index.tsx
@@ -1,9 +1,8 @@
 import styles from "./SignInPage.module.scss";
 import Link from "next/link";
 import Image from "next/image";
-import { SnsSign } from "@/components";
+import { SnsSign, Input } from "@/components/sign";
 import Head from "next/head";
-import Input from "@/components/Input/Input";
 
 function SignInPage() {
   return (
